% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LRSTbase.R
\name{gen.fun.dep}
\alias{gen.fun.dep}
\title{Title Generate Data for two outcomes.}
\usage{
gen.fun.dep(xm, xs, corln, Sigma)
}
\arguments{
\item{xm}{A mean matrix with element (i,j) giving the mean of outcome i and time j.}

\item{xs}{A s.d. matrix with element (i,j) giving the s.d. of outcome i and time j.}

\item{corln}{Correlation matrix between the outcomes.}

\item{Sigma}{Covariance Matrix for Random Effect}
}
\value{
Longitudinal Data with multiple outcomes
}
\description{
Title Generate Data for two outcomes.
}
\details{
This function is used inside a for-loop to simulate repeated measures per subject
}
\examples{
# Example: simulate 1 subject with 7 visits
library(MASS)
M <- -t(cbind(
  c(0, 0.738, 1.313, 3.109, 4.525, 5.864, 7.338),
  c(0, 0.668, 3.975, 5.931, 8.259, 11.989, 13.976)
))

S <- t(cbind(
  c(0, 4.79, 5.43, 6.54, 7.37, 8.15, 9.11),
  c(0, 10.27, 12.85, 14.95, 15.35, 16.87, 18.19)
))

Sigma <- matrix(c(3.477^2, 0, 0, 3.477^2), nrow = 2)
corr <- matrix(c(1, 0.5, 0.5, 1), nrow = 2)

gen.fun.dep(M, S, corr, Sigma)

}
