% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LRSTMulti.R
\name{lrst.MultiArm}
\alias{lrst.MultiArm}
\title{Longitudinal Rank Sum test for multi-arm Clinical Trial}
\usage{
lrst.MultiArm(X_c, Y)
}
\arguments{
\item{X_c}{Longitudinal Data for placebo of dimension \eqn{(n_x, K, T)}, where \eqn{n_x} is the number of
patients in the placebo group, K is the number of outcomes, and T is the number of time points}

\item{Y}{A list of length A, where A is the number of arms. Each element of Y is a longitudinal
data for the corresponding treatment arm of dimension \eqn{(n_{y_i}, K, T)}, where \eqn{n_{y_i}} is the number of
patients in the \eqn{i^{th}} treatment arm, K is the number of outcomes, and T is the number of time points.}
}
\value{
A list containing:
\item{p}{The computed p-value}
\item{v}{Proportion of times  higher dose was selected}
\item{T}{Final test Statistic}
}
\description{
Performs a non-parametric U-statistic test to compare multiple treatment arms against a placebo
using multivariate longitudinal outcome data. Returns the test statistic, p-value, and the selected arm.
}
\examples{
library(MASS)
xm = -t(cbind(c(0, 0.738, 1.313, 3.109, 4.525, 5.864, 7.338),
          c(0, 0.668, 3.975, 5.931, 8.259, 11.989, 13.976)))
xs = t(cbind(c(0, 4.79, 5.43, 6.54, 7.37, 8.15, 9.11),
          c(0, 10.27, 12.85, 14.95, 15.35, 16.87, 18.19)))
N = 100 # Total Number of Patients
nx = 2/5*N # Niumber of patients in Placebo
ny = 3/5*N # Number of patients in Treatment
K = 2 # Number of Outcomes
T = 6 # Number of visits other than baseline
sigma_u <- c(3.477, 7.232) # Random Effect Standard Deviation for Outcomes
corr <- matrix(c(1,0.5, 0.5, 1),2,2)  # Correlation between the outcomes
sigma = diag(sigma_u) \%*\% corr \%*\% diag(sigma_u)
alpha=1
beta=1
effect_size_1 = 2.21*alpha/6
effect_size_2 = 5.38*beta/6
ym = xm + t(cbind(c(0,(rep(effect_size_1,6))), c(0,(rep(effect_size_2,6)))))
X_placebo = array(NA, dim = c(nx, 2, 7))
X_dose_1 = array(NA, dim = c(ny, 2, 7))
X_dose_2 = array(NA, dim = c(ny, 2, 7))

for(i in 1:nx){
  X_placebo[i,,] = gen.fun.dep(xm, xs, corr, sigma)
}
for(i in 1:ny){
  X_dose_1[i,,] = gen.fun.dep(ym, xs, corr, sigma)
}
for(i in 1:ny){
  X_dose_2[i,,] = gen.fun.dep(ym, xs, corr, sigma)
}
X_c = X_placebo
Y = vector(length = 2, mode = "list")
Y[[1]] = X_dose_1
Y[[2]] = X_dose_2
lrst.MultiArm(X_c, Y)
}
\references{
Ghosh, D., Luo, S. (2024). A non-parametric U-statistic testing approach for multi-arm clinical trials with multivariate longitudinal data. arXiv preprint arXiv:2408.10149.
}
