% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Power.R
\name{sampSize}
\alias{sampSize}
\title{Title Minimum Sample Size required to obtain a given power}
\usage{
sampSize(pi, xm, ym, alpha, C, D, lambda)
}
\arguments{
\item{pi}{Required Power}

\item{xm}{Placebo mean matrix of size K * T.}

\item{ym}{Treatment mean matrix of size K * T}

\item{alpha}{Size of Test}

\item{C}{Theoretical C matrix}

\item{D}{Theoretical D matrix}

\item{lambda}{nx/ny where nx and ny are number of patients in the placebo and treatment groups respectively}
}
\value{
Minium Sample Size
}
\description{
Title Minimum Sample Size required to obtain a given power
}
