% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Power.R
\name{theoretical_power}
\alias{theoretical_power}
\title{Title Theoretical Power when C and D are known}
\usage{
theoretical_power(alpha = 0.05, effect_size, nx, ny, K, T, C, D)
}
\arguments{
\item{alpha}{Size of the Test}

\item{effect_size}{Theoretical Effect Size}

\item{nx}{Number of patients in the plcabo group}

\item{ny}{Number of patients in the treatment group}

\item{K}{Number of outcomes}

\item{T}{Number of Time Points}

\item{C}{Theoretical C matrix}

\item{D}{Theoretica; D matrix}
}
\value{
Theoretical Power
}
\description{
Title Theoretical Power when C and D are known
}
\examples{
N = 500 # Total Number of Patients
nx = 2/5*N # Niumber of patients in Placebo
ny = 3/5*N # Number of patients in Treatment
K = 2 # Number of Outcomes
T = 6 # Number of visits other than baseline
C = matrix(rnorm(T^2), nrow = T)
D = matrix(rnorm(T^2), nrow = T)
theoretical_power(alpha = 0.05, effect_size = 1, nx, ny, K, T+1, C, D)
}
\references{
Ghosh, D., Xu, X., Luo, S. (2025). Power and sample size calculation for multivariate longitudinal trials using the longitudinal rank sum test. arXiv preprint arXiv:2502.07152.
}
