% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Power.R
\name{theta.bar.theo}
\alias{theta.bar.theo}
\title{Title Theoretical Relative Treatment Effect Size Under Gaussianity Assumption}
\usage{
theta.bar.theo(xm, ym, xs, n.iter = 1000)
}
\arguments{
\item{xm}{Placebo Mean}

\item{ym}{Treatment Mean}

\item{xs}{Placebo Standard Deviation}

\item{n.iter}{Number of iteartions}
}
\description{
Title Theoretical Relative Treatment Effect Size Under Gaussianity Assumption
}
\examples{
xm = -t(cbind(c(0, 0.738, 1.313, 3.109, 4.525, 5.864, 7.338),
              c(0, 0.668, 3.975, 5.931, 8.259, 11.989, 13.976)))
xs = t(cbind(c(0, 4.79, 5.43, 6.54, 7.37, 8.15, 9.11),
              c(0, 10.27, 12.85, 14.95, 15.35, 16.87, 18.19)))
alpha=1
beta=1
effect_size_1 = 2.21*alpha/6
effect_size_2 = 5.38*beta/6
ym = xm + t(cbind(c(0,(rep(effect_size_1,6))), c(0,(rep(effect_size_2,6)))))
theta.bar.theo(xm, ym, xs, n.iter=10)
}
